refactor(md): restructure into domain-based modules
Split monolithic files into focused modules:
- processor/ (pipeline factory, sanitize schema, serialize, TOC, version)
- components/ (Markdown, RenderedMarkdown, default components)
- hooks/ (useMergedComponents shared hook)
- utils/ (renderHast utility)

Bugs fixed:
- Sanitizer wildcard: style, className, id, aria-hidden now correctly
  assigned back to schema (was orphaned in local array)
- style removed from wildcard, allowed only on specific MathML/SVG tags
- hr, steps, step missing type annotations added
- img unsafe type casts replaced with safe Number() conversion
- strip-markdown removed as dead code (TOC extracts clean text from HAST)

Enhancements:
- RENDERING_VERSION bumped to 6, RENDERING_CONFIG_ID added
- Additive @xbrk/md/react subpath export for component-only imports
