+++ packages/md/src/components/index.ts
@@ -1,0 +1,4 @@
+export type { Components } from 'hast-util-to-jsx-runtime';
+export { components } from './default-components';
+export { Markdown } from './markdown';
+export { RenderedMarkdown } from './rendered-markdown';
