+++ packages/md/src/processor/version.ts
@@ -1,0 +1,3 @@
+export const RENDERING_VERSION = 8;
+
+export const RENDERING_CONFIG_ID = 'remark-gfm+math+emoji+sanitize-no-clobber+rehype-katex+slug+autolink+shiki-v1';
